Remove make firmware_install (linux 4.14)
The in-tree firmware apparently wasn't used since 2013 and has finally been dropped in linux 4.14. `${MAKE} firmware_install` is not anymore present in linux 4.14 and buildkernel will fail to build the kernel. As the in-tree firmware hasn't been used for a long time it should be safe to remove ` ${MAKE} firmware_install` altogether.